[wip] generate metadata container
